Updated uninstall.php to check if user wants settings saved or not

+++ pinterest-widgets/uninstall.php
@@ -11,5 +11,16 @@ if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 	exit;
 }
 
-// Delete general settings option
-delete_option( 'pw_settings_general' );
+global $pw_options;
+
+$general = get_option( 'pw_settings_general' );
+
+if( ! empty( $general['uninstall_save_settings'] ) ) {
+	// DO Nothing since they want to save their settings on uninstall
+} else {
+	// Delete options because they don't want them saved
+	
+	// Delete general settings option
+	delete_option( 'pw_settings_general' );
+}
+
